requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
python-multipart==0.0.9
jinja2==3.1.4
email-validator==2.2.0
python-dotenv==1.0.1
pytesseract==0.3.10
Pillow==10.2.0

